// src/UI-components/organisms/OrderCardInfo/OrderCardInfo.type.ts
export interface Data {
    id: number
    cardName: string
    paymentSystem: string
    cashback: number
    coBrand: string
    isVirtual: boolean
    premiumStatus: string
    servicePrice: number
    productPrice: number
    currencyCode: string
    isActive: boolean
    cardDuration: number
}